refactor: move side-effects of selecting a template up to the parent
this allows the OrganizationTemplates component to not need to know about
the context it is being used on (e.g. the CreateProject or Organization screens)
