+++ src/client/screens/OrganizationScreen.tsx
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import { useEffect, useState } from "react";
 import { connect } from "react-redux";
-import { useParams, Link as RouterLink } from "react-router-dom";
+import { useParams, Link as RouterLink, useHistory } from "react-router-dom";
 import { Box, Button, Flex, Heading, Image, jsx, Spinner, Text, Link } from "theme-ui";
 import { formatDate } from "../functions";
 
@@ -17,7 +17,14 @@ import { UserState } from "../reducers/user";
 import { OrganizationProjectsState } from "../reducers/organizationProjects";
 import store from "../store";
 
-import { CreateProjectData, IOrganization, IUser, ProjectNest } from "../../shared/entities";
+import {
+  CreateProjectData,
+  IOrganization,
+  IUser,
+  ProjectNest,
+  IProjectTemplate,
+  IProject
+} from "../../shared/entities";
 
 import Icon from "../components/Icon";
 import JoinOrganizationModal from "../components/JoinOrganizationModal";
@@ -27,6 +34,7 @@ import FeaturedProjectCard from "../components/FeaturedProjectCard";
 import OrganizationTemplates from "../components/OrganizationTemplates";
 
 import PageNotFoundScreen from "./PageNotFoundScreen";
+import { createProject } from "../api";
 
 interface StateProps {
   readonly organization: OrganizationState;
@@ -141,6 +149,29 @@ const OrganizationScreen = ({ organization, organizationProjects, user }: StateP
 
   const userIsVerified = "resource" in user && user.resource && user.resource.isEmailVerified;
   const orgAdminUrl = `/o/${organizationSlug}/admin`;
+  const history = useHistory();
+
+  function createProjectFromTemplate(data: CreateProjectData) {
+    void createProject(data).then((project: IProject) => history.push(`/projects/${project.id}`));
+  }
+
+  function setupProjectFromTemplate(template: IProjectTemplate) {
+    const { id, name, regionConfig, numberOfDistricts, districtsDefinition, chamber } = template;
+    const data: CreateProjectData = {
+      name,
+      regionConfig,
+      numberOfDistricts,
+      districtsDefinition,
+      chamber,
+      projectTemplate: { id }
+    };
+    if (userInOrg) {
+      createProjectFromTemplate(data);
+    } else {
+      setProjectTemplate(data);
+      store.dispatch(showCopyMapModal(true));
+    }
+  }
 
   const featuredProjects =
     "resource" in organizationProjects.featuredProjects
@@ -307,7 +338,7 @@ const OrganizationScreen = ({ organization, organizationProjects, user }: StateP
                 <OrganizationTemplates
                   user={"resource" in user ? user.resource : undefined}
                   organization={organization.resource}
-                  setTemplate={setProjectTemplate}
+                  setTemplate={setupProjectFromTemplate}
                 />
               )}
             </Box>
